Add LED control and timeouts
- When in the DATA state, the LEDs are set based on the received packet.
- The DISCOVERY state will timeout and return to IDLE if an ACK is not
received from the server in time.
- If a discovery message is received while in the DATA state, we return
to IDLE in order to re-establish connection with the server. This
prevents the client from hanging when the server is reset.
- The UDP port is now fixed throughout the state machine, which enables
discovery messages to still be received while in the DATA state. This
may open up a security concern (e.g. having randomized ports is safer),
but for this use case is an acceptable risk.